+++ scrapyd/webservice.py
@@ -1,4 +1,5 @@
 import functools
+import re
 import sys
 import traceback
 import uuid
@@ -7,29 +8,32 @@ from copy import copy
 from io import BytesIO
 
 from twisted.python import log
-from twisted.web import http
-from twisted.web.error import Error
+from twisted.web import error, http
 
 from scrapyd.exceptions import MissingRequiredArgument
 from scrapyd.jobstorage import job_items_url, job_log_url
-from scrapyd.utils import JsonResource, UtilsCache, check_disallowed_characters, get_spider_list, native_stringify_dict
+from scrapyd.utils import JsonResource, UtilsCache, get_spider_list, native_stringify_dict
+
+# Allow the same characters as Setuptools. Setuptools uses underscores for replacement in filenames, and hyphens
+# for replacement in distributions. Quotes and spaces are allowed in case a config sets `project = "my project"`.
+# https://github.com/pypa/setuptools/blob/e304e4d/setuptools/_distutils/command/install_egg_info.py#L74
+DISALLOWED_CHARACTERS = re.compile(b"""[^A-Za-z0-9."'\\s_-]""")
 
 
-def with_safe_project_name(func):
-    @functools.wraps(func)
-    def wrapper(resource, txrequest):
-        project_name = txrequest.args.pop(b'project', [b''])[0].decode()
-        msg = "Project name is required and must be a valid string. "
-        msg += f"Project name '{project_name}' is not a valid project name."
-        msg = msg.encode()
-        if not project_name:
-            raise Error(code=400, message=msg)
+def with_project(required=True):
+    def decorator(func):
+        @functools.wraps(func)
+        def wrapper(resource, txrequest):
+            project = txrequest.args.pop(b'project', [b''])[0]
+            if required and not project:
+                raise error.Error(code=http.BAD_REQUEST, message=b"'project' parameter is required")
+            if DISALLOWED_CHARACTERS.search(project):
+                raise error.Error(code=http.BAD_REQUEST, message=b"project '%b' is invalid" % project)
+            return func(resource, txrequest, project.decode())
 
-        if not check_disallowed_characters(project_name):
-            raise Error(code=400, message=msg)
-        return func(resource, txrequest, project_name)
+        return wrapper
 
-    return wrapper
+    return decorator
 
 
 def _get_required_param(args, param):
@@ -55,7 +59,7 @@ class WsResource(JsonResource):
         try:
             return JsonResource.render(self, txrequest).encode('utf-8')
         except Exception as e:
-            if isinstance(e, Error):
+            if isinstance(e, error.Error):
                 txrequest.setResponseCode(e.args[0])
             if self.root.debug:
                 return traceback.format_exc().encode('utf-8')
@@ -96,7 +100,7 @@ class DaemonStatus(WsResource):
 
 
 class Schedule(WsResource):
-    @with_safe_project_name
+    @with_project()
     def render_POST(self, txrequest, project):
         args = native_stringify_dict(copy(txrequest.args), keys_only=False)
         settings = args.pop('setting', [])
@@ -116,9 +120,9 @@ class Schedule(WsResource):
 
 
 class Cancel(WsResource):
-    def render_POST(self, txrequest):
+    @with_project()
+    def render_POST(self, txrequest, project):
         args = {k: v[0] for k, v in native_stringify_dict(copy(txrequest.args), keys_only=False).items()}
-        project = _get_required_param(args, 'project')
         jobid = _get_required_param(args, 'job')
         # Instead of os.name, use sys.platform, which disambiguates Cygwin, which implements SIGINT not SIGBREAK.
         # https://cygwin.com/cygwin-ug-net/kill.html
@@ -141,7 +145,7 @@ class Cancel(WsResource):
 
 
 class AddVersion(WsResource):
-    @with_safe_project_name
+    @with_project()
     def render_POST(self, txrequest, project):
         egg = _pop_required_param(txrequest.args, b'egg')[0]
         if not zipfile.is_zipfile(BytesIO(egg)):
@@ -164,14 +168,14 @@ class ListProjects(WsResource):
 
 
 class ListVersions(WsResource):
-    @with_safe_project_name
+    @with_project()
     def render_GET(self, txrequest, project):
         versions = self.root.eggstorage.list(project)
         return {"node_name": self.root.nodename, "status": "ok", "versions": versions}
 
 
 class ListSpiders(WsResource):
-    @with_safe_project_name
+    @with_project()
     def render_GET(self, txrequest, project):
         args = native_stringify_dict(copy(txrequest.args), keys_only=False)
         version = args.get('_version', [''])[0]
@@ -180,10 +184,10 @@ class ListSpiders(WsResource):
 
 
 class Status(WsResource):
-    def render_GET(self, txrequest):
+    @with_project(required=False)
+    def render_GET(self, txrequest, project):
         args = native_stringify_dict(copy(txrequest.args), keys_only=False)
         job = _get_required_param(args, 'job')[0]
-        project = args.get('project', [None])[0]
 
         spiders = self.root.launcher.processes.values()
         queues = self.root.poller.queues
@@ -191,16 +195,16 @@ class Status(WsResource):
         result = {"node_name": self.root.nodename, "status": "ok", "currstate": "unknown"}
 
         for s in self.root.launcher.finished:
-            if (project is None or s.project == project) and s.job == job:
+            if (not project or s.project == project) and s.job == job:
                 result["currstate"] = "finished"
                 return result
 
         for s in spiders:
-            if (project is None or s.project == project) and s.job == job:
+            if (not project or s.project == project) and s.job == job:
                 result["currstate"] = "running"
                 return result
 
-        for qname in (queues if project is None else [project]):
+        for qname in ([project] if project else queues):
             for x in queues[qname].list():
                 if x["_job"] == job:
                     result["currstate"] = "pending"
@@ -210,16 +214,14 @@ class Status(WsResource):
 
 
 class ListJobs(WsResource):
-    def render_GET(self, txrequest):
-        args = native_stringify_dict(copy(txrequest.args), keys_only=False)
-        project = args.get('project', [None])[0]
-
+    @with_project(required=False)
+    def render_GET(self, txrequest, project):
         spiders = self.root.launcher.processes.values()
         queues = self.root.poller.queues
 
         pending = [
             {"project": qname, "spider": x["name"], "id": x["_job"]}
-            for qname in (queues if project is None else [project])
+            for qname in ([project] if project else queues)
             for x in queues[qname].list()
         ]
         running = [
@@ -231,7 +233,7 @@ class ListJobs(WsResource):
                 "start_time": str(s.start_time),
             }
             for s in spiders
-            if project is None or s.project == project
+            if not project or s.project == project
         ]
         finished = [
             {
@@ -244,7 +246,7 @@ class ListJobs(WsResource):
                 "items_url": job_items_url(s),
             }
             for s in self.root.launcher.finished
-            if project is None or s.project == project
+            if not project or s.project == project
         ]
 
         return {"node_name": self.root.nodename, "status": "ok",
@@ -252,7 +254,7 @@ class ListJobs(WsResource):
 
 
 class DeleteProject(WsResource):
-    @with_safe_project_name
+    @with_project()
     def render_POST(self, txrequest, project):
         self._delete_version(project)
         UtilsCache.invalid_cache(project)
@@ -264,7 +266,7 @@ class DeleteProject(WsResource):
 
 
 class DeleteVersion(DeleteProject):
-    @with_safe_project_name
+    @with_project()
     def render_POST(self, txrequest, project):
         args = native_stringify_dict(copy(txrequest.args), keys_only=False)
         version = _get_required_param(args, 'version')[0]